12/08

+++ src/main/java/web/model/User.java
@@ -1,6 +1,5 @@
 package web.model;
 
-
 import javax.persistence.*;
 
 @Entity
@@ -14,12 +13,12 @@ public class User {
     @Column (name = "lastname")
     private String lastName;
     @Column (name = "age")
-    private Byte age;
+    private int age;
 
     public User() {
     }
 
-    public User(String name, String lastName, Byte age) {
+    public User(String name, String lastName, int age) {
         this.name = name;
         this.lastName = lastName;
         this.age = age;
@@ -49,11 +48,11 @@ public class User {
         this.lastName = lastName;
     }
 
-    public Byte getAge() {
+    public int getAge() {
         return age;
     }
 
-    public void setAge(Byte age) {
+    public void setAge(int age) {
         this.age = age;
     }
